swig[RF] major swignifit refactor
The part of swignifit that is responsible for interfacing with Psi++ has been
renamed. The new name for the whole thing is 'swignift_raw'. The idea is that
'swignifit' is a package that contains both the 'raw' code to interface with
Psi++, as well as the 'interface' code that lies ontop of the raw code and
provides convenience functions/classes for the user, to give the following
package layout:

swignifit.interface
swignifit.swignifit_raw

Though I am not completely satisfied with the name of 'swignifit.interface', it
just causes ambiguity.

The swig interface file, 'swignifit.i' has been renamed to reflect this change.
It will now generate two files, swignifit_raw.py and swignifit_raw.cxx. The C++
source file is compiled with the Psi++ code to yield the _swignifit_raw.o shared
library, which now lives in the swignifit package instead of in the
root package. As a result of this you now need to something along the lines of
the following to test code without installing it:

cd .. ; PYTHONPATH=psignifit/build/lib.linux-x86_64-2.5 ipython

The problem is that it does not seem possible to remove the $PWD from the
PYTHONPATH. In the case of the compiled extension the first place the
interpreter looks is NOT in the build (psignifit/build/lib.*/swignifit)
directory, but instaed in the source psignifit/swignifit directory, which never
contains any compiled extensions, and as a result raises an import error.

The Makefile, the setup_basic.py and the tests have been adapted to reflect
these changes.