BF: increase speed, remove unnecessary calls to getstart
Profiling and user reports showed that there was a problem in
psignifit-bootstrap command line interface. After inspecting the call graph we
discovered that the main bottleneck were unnecessary calls to getstart by
jackknife via optimize. The getstart is inherently quite slow but gains its
popularity because it needs to be called only once. The problem was that
getstart was being called multiple times -- once for each jackknifed dataset. We
optimized this by obtaining starting value once at the beginning of the jackknife
routine and then reusing these for each jackknifed dataset.

Initial profiling for a test dataset showed 50% speed increase.