RF: make tests against pfit2 more lenient
After optimizing our optimizer we obtain a deviance value for the solution that
is within 1e-2 of the original pfit2 solution and not 0.2 different.

pfit2 deviance:  2.08172108159
pfit3 pre-optimizer optimization deviance 2.21098992844
pfit3 post-optimizer optimization solution 2.08331854399

Thus we make the checks for the solution and for the deviance value more
lenient, but keep the original values given by pfit2, since these are the ones
we are checking against. The devinace is still a little worse... But not as bad
as before. When taking a closer look at the solution we see that it does not
quite reach the bottom of the minimum, but it does get close enough for all
practical purposes.